Split IteratorUtil up into 2 non-API classes
* Moved methods not useful to users to non-API classes
* Created IterConfigUtil and SystemIteratorUtil for internal use
* Cleaned up loadIterators methods and created IterLoad class to assist
